requirements.txt
discord.py==2.4.0
cloudscraper==1.2.71
requests==2.32.3